pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "dm-lab-benchmark-suite"
version = "0.1.0"
description = "Local-first reproducible data mining benchmark suite"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "DM-Lab Team" }]
dependencies = [
  "numpy>=2.1",
  "pandas>=2.2",
  "scikit-learn>=1.6",
  "mlxtend>=0.23",
  "imbalanced-learn>=0.13",
  "matplotlib>=3.10",
  "seaborn>=0.13",
  "streamlit>=1.42",
  "PyYAML>=6.0",
  "requests>=2.32"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.11"
]

[project.scripts]
dm-lab = "dm_lab.cli:entrypoint"

[tool.pytest.ini_options]
pythonpath = ["src", "."]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.hatch.build.targets.wheel]
packages = ["src/dm_lab"]
